Add files via upload
This work proposes a statistical model for crossover trials with multiple skewed responses measured in each period. A 3 $\times$ 3 crossover trial data where different doses of a drug were administered to subjects with a history of seasonal asthma rhinitis to grass pollen is used for motivation. In each period, gene expression values for ten genes were measured from each subject. It considers a linear mixed effect model with skew normally distributed random effect or random error term to model the asymmetric responses in the crossover trials. The article examines cases (i) when a random effect follows a skew-normal distribution, as well as (ii) when a random error follows a skew-normal distribution. The EM algorithm is used in both cases to compute maximum likelihood estimates of parameters. Simulations and crossover data from the gene expression study illustrate the proposed approach.